requirements.txt
opencv-python==4.6.0.66
Pillow==9.3.0
numpy==1.20
Flask==2.2.2
Flask-Cors==3.0.10
protobuf==3.17.3
gunicorn